Fix extension function scope
